Lat/Lon decimal fix, remove kml before render, updated index and infowindow

+++ db/migrate/20120108164400_add_sidewalk_attributes.rb
@@ -1,8 +1,8 @@
 class AddSidewalkAttributes < ActiveRecord::Migration
   def change
     change_table :chicagosidewalks do |t|
-      t.integer :lat, :default => nil, :precision => 16, :scale => 14
-      t.integer :lon, :default => nil, :precision => 16, :scale => 14
+      t.decimal :lat, :default => nil, :precision => 16, :scale => 14
+      t.decimal :lon, :default => nil, :precision => 16, :scale => 14
       t.boolean :cleared
       t.boolean :need_help
       t.string :zipcode
